refactor project as a package
