+++ DeepBach/datasets/custom_dataset/.README
@@ -1,0 +1,1 @@
+This directory is meant to contain preprocessed user-defined corpora.
