+++ DeepBach/datasets/raw_dataset/.README
@@ -1,0 +1,1 @@
+This directory is meant to contain preprocessed J.S. Bach chorales.
