--- datasets/custom_dataset/.README
@@ -1,1 +1,0 @@
-This directory is meant to contain preprocessed user-defined corpora.
