--- datasets/raw_dataset/.README
@@ -1,1 +1,0 @@
-This directory is meant to contain preprocessed J.S. Bach chorales.
