+++ plugin_flask_server.py
@@ -1,18 +1,16 @@
 import os
 import pickle
 import tempfile
-import sys
 from glob import glob
 
-from data_utils import START_SYMBOL, END_SYMBOL, all_features, \
-    all_metadatas, indexed_chorale_to_score, chorale_to_inputs, BACH_DATASET
-from deepBach import load_models
+import numpy as np
 from flask import Flask, request, make_response, jsonify
-
 from music21 import musicxml, converter
 from tqdm import tqdm
 
-import numpy as np
+from DeepBach.data_utils import START_SYMBOL, END_SYMBOL, all_features, \
+    all_metadatas, indexed_chorale_to_score, chorale_to_inputs, BACH_DATASET
+from DeepBach.model_manager import load_models
 
 UPLOAD_FOLDER = '/tmp'
 ALLOWED_EXTENSIONS = {'xml', 'mxl', 'mid', 'midi'}
@@ -56,12 +54,15 @@ def parallel_gibbs_server(models=None,
 
     for expert_index in range(num_voices):
         # Add start and end symbol
-        seq[:timesteps, expert_index] = [note2indexes[expert_index][START_SYMBOL]] * timesteps
-        seq[-timesteps:, expert_index] = [note2indexes[expert_index][END_SYMBOL]] * timesteps
+        seq[:timesteps, expert_index] = [note2indexes[expert_index][
+                                             START_SYMBOL]] * timesteps
+        seq[-timesteps:, expert_index] = [note2indexes[expert_index][
+                                              END_SYMBOL]] * timesteps
     for expert_index in range(start_voice_index, end_voice_index + 1):
         # Randomize selected zone
-        seq[timesteps + start_tick: timesteps + end_tick, expert_index] = np.random.randint(num_pitches[expert_index],
-                                                                                            size=end_tick - start_tick)
+        seq[timesteps + start_tick: timesteps + end_tick,
+        expert_index] = np.random.randint(num_pitches[expert_index],
+                                          size=end_tick - start_tick)
 
     if chorale_metas is not None:
         # chorale_metas is a list
@@ -81,7 +82,8 @@ def parallel_gibbs_server(models=None,
     # Main loop
     for iteration in tqdm(range(num_iterations)):
 
-        temperature = max(min_temperature, temperature * discount_factor)  # Simulated annealing
+        temperature = max(min_temperature,
+                          temperature * discount_factor)  # Simulated annealing
 
         time_indexes = {}
         probas = {}
@@ -92,17 +94,20 @@ def parallel_gibbs_server(models=None,
             time_indexes[voice_index] = []
 
             for batch_index in range(batch_size_per_voice):
-                time_index = np.random.randint(timesteps + start_tick, timesteps + end_tick)
+                time_index = np.random.randint(timesteps + start_tick,
+                                               timesteps + end_tick)
                 time_indexes[voice_index].append(time_index)
 
                 (left_feature,
                  central_feature,
                  right_feature,
-                 label) = all_features(seq, voice_index, time_index, timesteps, num_pitches, num_voices)
+                 label) = all_features(seq, voice_index, time_index, timesteps,
+                                       num_pitches, num_voices)
 
-                left_metas, central_metas, right_metas = all_metadatas(chorale_metadatas=extended_chorale_metas,
-                                                                       metadatas=metadatas,
-                                                                       time_index=time_index, timesteps=timesteps)
+                left_metas, central_metas, right_metas = all_metadatas(
+                    chorale_metadatas=extended_chorale_metas,
+                    metadatas=metadatas,
+                    time_index=time_index, timesteps=timesteps)
 
                 input_features = {'left_features': left_feature[:, :],
                                   'central_features': central_feature[:],
@@ -116,12 +121,15 @@ def parallel_gibbs_server(models=None,
                 batch_input_features.append(input_features)
 
             # convert input_features
-            batch_input_features = {key: np.array([input_features[key] for input_features in batch_input_features])
+            batch_input_features = {key: np.array(
+                [input_features[key] for input_features in
+                 batch_input_features])
                                     for key in batch_input_features[0].keys()
                                     }
             # make all estimations
-            probas[voice_index] = models[voice_index].predict(batch_input_features,
-                                                              batch_size=batch_size_per_voice)
+            probas[voice_index] = models[voice_index].predict(
+                batch_input_features,
+                batch_size=batch_size_per_voice)
             if not parallel_updates:
                 # update
                 for batch_index in range(batch_size_per_voice):
@@ -129,12 +137,14 @@ def parallel_gibbs_server(models=None,
 
                     # use temperature
                     probas_pitch = np.log(probas_pitch) / temperature
-                    probas_pitch = np.exp(probas_pitch) / np.sum(np.exp(probas_pitch)) - 1e-7
+                    probas_pitch = np.exp(probas_pitch) / np.sum(
+                        np.exp(probas_pitch)) - 1e-7
 
                     # pitch can include slur_symbol
                     pitch = np.argmax(np.random.multinomial(1, probas_pitch))
 
-                    seq[time_indexes[voice_index][batch_index], voice_index] = pitch
+                    seq[time_indexes[voice_index][
+                            batch_index], voice_index] = pitch
 
         if parallel_updates:
             # update
@@ -144,12 +154,14 @@ def parallel_gibbs_server(models=None,
 
                     # use temperature
                     probas_pitch = np.log(probas_pitch) / temperature
-                    probas_pitch = np.exp(probas_pitch) / np.sum(np.exp(probas_pitch)) - 1e-7
+                    probas_pitch = np.exp(probas_pitch) / np.sum(
+                        np.exp(probas_pitch)) - 1e-7
 
                     # pitch can include slur_symbol
                     pitch = np.argmax(np.random.multinomial(1, probas_pitch))
 
-                    seq[time_indexes[voice_index][batch_index], voice_index] = pitch
+                    seq[time_indexes[voice_index][
+                            batch_index], voice_index] = pitch
 
     return seq[timesteps:-timesteps, :]
 
@@ -166,19 +178,23 @@ if not os.path.exists(pickled_dataset):
     raise NotImplementedError
 
 # load dataset
-X, X_metadatas, voice_ids, index2notes, note2indexes, metadatas = pickle.load(open(pickled_dataset, 'rb'))
+X, X_metadatas, voice_ids, index2notes, note2indexes, metadatas = pickle.load(
+    open(pickled_dataset, 'rb'))
 
 num_voices = len(voice_ids)
 num_pitches = list(map(len, index2notes))
 
 # get model names present in folder models/
 models_list = glob('models/*.yaml')
-models_list = list(set(map(lambda name: '_'.join(name.split('_')[:-1]).split('/')[-1], models_list)))
+models_list = list(set(
+    map(lambda name: '_'.join(name.split('_')[:-1]).split('/')[-1],
+        models_list)))
 
 # model_name = 'deepbach'
 model_name = 'skip_new'
 
-assert os.path.exists('models/' + model_name + '_' + str(num_voices - 1) + '.yaml')
+assert os.path.exists(
+    'models/' + model_name + '_' + str(num_voices - 1) + '.yaml')
 
 # load models
 models = load_models(model_name, num_voices=num_voices)
@@ -216,15 +232,18 @@ def compose():
         # make chorale time major
         input_chorale = np.transpose(input_chorale, axes=(1, 0))
         NUM_MIDI_TICKS_IN_SIXTEENTH_NOTE = 120
-        start_tick_selection = int(float(request.form['start_tick']) / NUM_MIDI_TICKS_IN_SIXTEENTH_NOTE)
-        end_tick_selection = int(float(request.form['end_tick']) / NUM_MIDI_TICKS_IN_SIXTEENTH_NOTE)
+        start_tick_selection = int(float(
+            request.form['start_tick']) / NUM_MIDI_TICKS_IN_SIXTEENTH_NOTE)
+        end_tick_selection = int(
+            float(request.form['end_tick']) / NUM_MIDI_TICKS_IN_SIXTEENTH_NOTE)
         # if no selection REGENERATE and set chorale length
         if start_tick_selection == 0 and end_tick_selection == 0:
             chorale_length = len(chorale_metas[0])
             # randomize
-            input_chorale = np.array([np.random.randint(num_pitches[expert_index],
-                                                        size=(chorale_length,))
-                                      for expert_index in range(num_voices)])
+            input_chorale = np.array(
+                [np.random.randint(num_pitches[expert_index],
+                                   size=(chorale_length,))
+                 for expert_index in range(num_voices)])
             input_chorale = np.transpose(input_chorale, axes=(1, 0))
             end_tick_selection = chorale_length
             start_voice_index = 0
@@ -241,9 +260,8 @@ def compose():
         else:
             batch_size_per_voice = 16
 
-        num_iterations = max(int(num_iterations // batch_size_per_voice // num_voices), 5)
-
-
+        num_iterations = max(
+            int(num_iterations // batch_size_per_voice // num_voices), 5)
 
         # --- Generate---
         output_chorale = parallel_gibbs_server(models=models,
@@ -262,9 +280,10 @@ def compose():
                                                metadatas=metadatas)
 
         # convert back to music21
-        output_chorale = indexed_chorale_to_score(np.transpose(output_chorale, axes=(1, 0)),
-                                                  pickled_dataset=pickled_dataset
-                                                  )
+        output_chorale = indexed_chorale_to_score(
+            np.transpose(output_chorale, axes=(1, 0)),
+            pickled_dataset=pickled_dataset
+            )
 
         # convert chorale to xml
         goe = musicxml.m21ToXml.GeneralObjectExporter(output_chorale)
@@ -289,7 +308,9 @@ def get_models():
     global models_list
     # recompute model names present in folder models/
     models_list = glob('models/*.yaml')
-    models_list = list(set(map(lambda name: '_'.join(name.split('_')[:-1]).split('/')[-1], models_list)))
+    models_list = list(set(
+        map(lambda name: '_'.join(name.split('_')[:-1]).split('/')[-1],
+            models_list)))
     return jsonify(models_list)
